backend/requirements.txt
django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
scikit-learn==1.6.1
pandas==2.2.3
xgboost==3.0.0
tensorflow==2.19.0
joblib==1.4.2
matplotlib==3.10.1
numpy==2.1.3
scipy==1.15.2 